+++ includes/core/Activation.php
@@ -26,16 +26,13 @@ class Activation{
         }
 
         $activation = $this->save_activation_key();
-        $activation_status = self::validate_saved_activation_key_from_anar();
+        $activation_status = self::validate_token();
 
         if ($activation && $activation_status) {
             $response = [
                 'success' => true,
                 'message' => 'توکن شما با موفقیت ثبت و تایید شد',
             ];
-
-            // update anar unpaid orders on wpdb to show alert
-            (new Payments())->count_unpaid_orders_count();
 
         } elseif ($activation_status) {
             $response = [
@@ -56,6 +53,8 @@ class Activation{
             if (isset($_POST['activation_code'])) {
                 $activation_code = sanitize_text_field($_POST['activation_code']);
                 $activation = update_option('_awca_activation_key', $activation_code);
+//                @todo change token key
+                update_option('_anar_token', $activation_code);
                 if ($activation) {
                     return true;
                 } else {
@@ -76,30 +75,47 @@ class Activation{
     }
 
 
-    public static function validate_saved_activation_key_from_anar()
+    public static function validate_token()
     {
 
         if(!self::get_saved_activation_key())
             return false;
 
         $tokenValidation = ApiDataHandler::tryGetAnarApiResponse("https://api.anar360.com/wp/auth/validate");
+
         if ($tokenValidation !== null) {
+            if(isset($tokenValidation->shopUrl)){
+                update_option('_anar_shop_url', $tokenValidation->shopUrl);
+            }
+            if(isset($tokenValidation->subscriptionPlan)){
+                update_option('_anar_subscription_plan', $tokenValidation->subscriptionPlan);
+            }
+            if(isset($tokenValidation->subscriptionRemaining)){
+                update_option('_anar_subscription_remaining', $tokenValidation->subscriptionRemaining);
+            }
+
             if (isset($tokenValidation->success) && $tokenValidation->success === true) {
-
-                if(isset($tokenValidation->shopUrl)){
-                    update_option('_anar_shop_url', $tokenValidation->shopUrl);
-                }
-                if(isset($tokenValidation->subscriptionPlan)){
-                    update_option('_anar_subscription_plan', $tokenValidation->subscriptionPlan);
-                }
-                if(isset($tokenValidation->subscriptionRemaining)){
-                    update_option('_anar_subscription_remaining', $tokenValidation->subscriptionRemaining);
-                }
-
-
+                update_option('_anar_token_validation', 'valid');
                 return true;
+            }else{
+                update_option('_anar_token_validation', 'invalid');
+                return false;
             }
         }
+        return false;
+    }
+
+    public static function is_active(){
+        $token_validation = get_option('_anar_token_validation', 'invalid');
+
+        if(!$token_validation){
+            return self::validate_token();
+        }elseif($token_validation === 'valid'){
+            return true;
+        }elseif($token_validation === 'invalid'){
+            return false;
+        }
+        awca_log('unknown error, anar token validation not saved, check Activation::is_active() method for more details');
         return false;
     }