+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # تغییرات
 تمام تغییرات قابل توجه این پروژه در این فایل مستند خواهد شد.
 
-## [0.3.13] - 1404/04/07
+## [0.3.14] - 1404/04/08
+- fix : _anar_sku of deprecated variable products not restored properly
+
+
+
+## [0.3.13] - 1404/04/08
 - feat : restore deprecation on syncOutdated
+- feat : show shopUrl and subscription data on setting page
+- feat : show anar product link based on shopUrl
 
 ## [0.3.11] - 1404/03/12
 - feat : slow Import method for low resources hosts